fix: password input
